requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-seed==0.3.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
Faker==33.3.0
flake8==7.1.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
uritemplate==4.1.1
